pyproject.toml
[tool.poetry]
name = "wows-stat-bot"
version = "0.1.0"
description = "my bot for game stat"
authors = ["Андрій <andrey.maybrodskyy@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "wows_stat_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.1.1"
python-dotenv = "^1.0.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
